requirements.txt
ImageTk
pillow
pytz
mysql-connector-python
configobj
httplib
http.client
prodict